+++ index.html
@@ -4,30 +4,24 @@
     <meta charset="UTF-8" />
     <link rel="favicon" href="favicon.ico" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <link
-      rel="stylesheet"
-      as="style"
-      crossorigin
-      href="https://cdn.jsdelivr.net/gh/orioncactus/pretendard@v1.3.6/dist/web/variable/pretendardvariable.css"
-    />
-    <link rel="stylesheet" href="//fonts.googleapis.com/earlyaccess/nanummyeongjo.css" />
+
+    <!-- 머테리얼 디자인 아이콘 -->
     <link
       rel="stylesheet"
       href="https://fonts.googleapis.com/css2?family=Material+Symbols+Outlined"
     />
-    <link rel="preconnect" href="https://fonts.googleapis.com" />
+
+    <!-- 이전 기수 api pre-connection. 리소스 로딩을 줄이려고 설정한 것 같다 -->
+    <!-- 백엔드가 api 주소를 바꾼다면 이후 지우거나 수정 -->
     <link rel="preconnect" href="https://moreturn.shop" crossorigin />
-    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
-    <link
-      href="https://fonts.googleapis.com/css2?family=Noto+Sans+KR&display=swap"
-      rel="stylesheet"
-    />
 
     <title>Moreturn 쉬운 부동산</title>
   </head>
   <body>
     <div id="root"></div>
     <script type="module" src="/src/main.tsx"></script>
+
+    <!-- 카카오 지도 api -->
     <script
       type="text/javascript"
       src="//dapi.kakao.com/v2/maps/sdk.js?appkey=697a355fbbe49be250eb554bcb37345c&libraries=services,clusterer"
