+++ package.json
@@ -23,7 +23,6 @@
     "emotion-reset": "^3.0.1",
     "eslint-plugin-react": "^7.32.2",
     "path": "^0.12.7",
-    "pretendard": "^1.3.6",
     "react": "^18.2.0",
     "react-daum-postcode": "^3.1.1",
     "react-dom": "^18.2.0",
