+++ src/components/HomeComp/HomeTitle/style.ts
@@ -1,5 +1,13 @@
 import styled from '@emotion/styled';
 
+interface FontProps {
+  color?: string;
+  size?: string;
+  weight?: 100 | 200 | 300 | 400 | 500 | 600 | 700 | 800 | 900;
+  lineHeight?: string;
+  align?: string;
+}
+
 export const HomeTitleContainer = styled.div`
   display: flex;
   flex-direction: column;
@@ -8,4 +16,12 @@ export const HomeTitleContainer = styled.div`
   .mainTitle {
     margin-bottom: 102px;
   }
+`;
+
+export const Title = styled.span<FontProps>`
+  font-weight: ${(props) => props.weight || 400};
+  font-size: ${(props) => props.size || '16px'};
+  color: ${(props) => props.color || '#000'};
+  line-height: ${(props) => props.lineHeight};
+  text-align: ${(props) => props.align};
 `;
